requirements.txt
fpdf2==2.7.4
pandas==2.0.0
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
scapy==2.5.0
numpy==1.24.3
pyattck==2.5.0
joblib==1.3.2
pillow==9.5.0